Refactor step
